requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib-resources==6.5.2
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
narwhals==1.45.0
numpy==2.0.2
packaging==25.0
pandas==2.3.0
pillow==11.3.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zipp==3.23.0
